+++ docs/release-notes.md
@@ -1,6 +1,10 @@
 ← [README](README.md)
 
 # Release notes
+## Upcoming release
+* For players:
+  * Added `set_farm_type` [console command](https://stardewvalleywiki.com/Modding:Console_commands#Console_commands) to change the current farm type.
+
 ## 3.12.8
 Released 18 October 2021 for Stardew Valley 1.5.4.